Add TTS so robot speech is said aloud
Since we are testing and running this node on Ubuntu 14.04, the TTS
client we are using is the built-in spd-say. An ambitious contributor
might add other TTS clients for other OSes, and pick which one to use
based on which OS we are running on. For now, though, on Ubuntu this
works to set the "robot is speaking" flag appropriately when the TTS
client is playing back sound.

While in this file, update format of comments to be capitalized with
periods.